fix: Keep rootProject.name to Gradle approved characters
Gradle wants the root project name to be from a small set of characters
that, specifically, doesn't include " ".

This isn't a problem when this is a standalone project, but it is a
problem if you try and include this code in another project as an
"included build" -- for example, if you're developing changes to the
library.

So rename to "android-connector" to match the module name.